Automated schema config, and reorganized endpoints
There is no need to configure the overblog_graphql extension in the app's config anymore.

In addition, the endpoints have been reorganized. Even though /repository is kept for BC,
the domain one is moved to the root. It also has an _repository field that gives access
to the repository's API, with contentTypes, sections...