Refactored how the schema is registered
A SchemaProvider interface, with a YamlSchemaProvider, detects defined schema files.
The Schema Provider is initialized by the bundle, and passed to the extension.
It is used by the extension to configure the graphql query and mutation configuration keys.

The Bundle is now used to store the paths to the different GraphQL elements through constants.
Doing so allows to make the schema provider testable, since the bundle provides the paths to
the other components (compiler pass, yaml schema provider).

Open questions:
- it may not be necessary to move the different path configs to the bundle. They may as well live in
  the extension