Added content._thumbnail
First iterates over fields, and returns the first ezimage or ezimageasset field. If none is found, tries with ezobjectrelation, and returns the first ezimage or ezimageasset field of the related item.

If an image field is found, generates the small variation of it, and returns it as a Thumbnail object with uri, width, height, alternativeText.