Fix EZP-30957: avoid exceptions on content types not in the generated schema
After a content type has been added, but before the schema has been generated, trying to list
a location's sub-items that contain one of those items will error out. The typeResolver of the
DomainContent interface will return for items of those types a name that doesn't exist in the schema.

To work around that, content items of types that don't exist in the schema are mapped to an "anonymous" type,
"UntypedContent".