Added the FieldDefinitionInputMapper interface
Implement on a Field Definition Mapper to generate a dynamic type for create and update mutations
on a given Field Type. The service must be tagged with ezplatform_graphql.field_definition_input_mapper,
with a fieldtype attribute set to the Field Type's identifier.

In version 2.0, the new interface will be merged into FieldDefinitionMapper,
and the service tag won't be needed anymore.